CASMUSER-2466: Convert a uan-img repo into uai-util
This is a clone of uan-img that is becoming uai-util.
The purpose of uai-util is to generate an rpm that could
be installed into an image to make that image capable
of becoming a UAI.

This is one step of several towards not building a "UAI"
docker image in the build pipeline.