CASMUAS-199: Add job launch capabilities to base UAI image
The biggest change here is to add the Slurm and PBS client
rpms so that this image could be used for basic job launch
testing. Also add some helpful commands like less, jq, curl,
etc...

In order to compile a useful application and run it with the
WLMs, a UAI image based on the compute node would still need to
be used. These additions at least provide an image that can
be used to validate UAS and job launch "out-of-the-box".